Redistribution of functions to new controller + added tests
