+++ itachallenge-user/src/main/java/com/itachallenge/user/service/UserService.java
@@ -17,6 +17,10 @@ public interface UserService {
 
     Mono<Boolean> deleteChallengeFromBookmarks(String userId, String challengeId);
 
+    Mono<Boolean> addToFavorites(UUID userUuid, UUID challengeUuid);
+
+    Mono<Boolean> deleteFromFavorites(UUID userId, UUID challengeUuid);
+
     Mono<Set<UUID>> getUserBookmarks(String userId);
     
     Mono<UserDocument> getUserById(String userId);
