+++ itachallenge-user/src/main/java/com/itachallenge/userinteraction/service/favorite/FavoriteService.java
@@ -7,5 +7,9 @@ import java.util.UUID;
 
 public interface FavoriteService {
 
-    public Mono<Set<UUID>> getUserFavorites(String userId);
+    Mono<Set<UUID>> getUserFavorites(String userId);
+    Mono<Boolean> addChallengeToFavorites(String userId, String challengeId);
+    Mono<Boolean> deleteChallengeFromFavorites(String userId, String challengeId);
+    Mono<Boolean> addToFavorites(UUID userUuid, UUID challengeUuid);
+    Mono<Boolean> deleteFromFavorites(UUID userId, UUID challengeUuid);
 }
