+++ itachallenge-user/src/main/java/com/itachallenge/userinteraction/service/favorite/FavoriteServiceImpl.java
@@ -3,20 +3,22 @@ package com.itachallenge.userinteraction.service.favorite;
 import com.itachallenge.user.exception.BadUUIDException;
 import com.itachallenge.userinteraction.document.favorite.FavoriteDocument;
 import com.itachallenge.userinteraction.repository.favorite.FavoriteRepository;
+import org.springframework.stereotype.Service;
 import reactor.core.publisher.Mono;
 
 import java.util.Set;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
+@Service
 public class FavoriteServiceImpl implements FavoriteService {
     private final FavoriteRepository favoriteRepository;
 
     public FavoriteServiceImpl(FavoriteRepository favoriteRepository) {
         this.favoriteRepository = favoriteRepository;
     }
-    private Mono<UUID> parseAndValidateUUID(String id) {
 
+    private Mono<UUID> parseAndValidateUUID(String id) {
         if (id == null || id.isEmpty()) {
             return Mono.error(new BadUUIDException("Invalid ID format"));
         }
@@ -28,7 +30,7 @@ public class FavoriteServiceImpl implements FavoriteService {
         }
     }
 
-@Override
+    @Override
     public Mono<Set<UUID>> getUserFavorites(String userId) {
         return parseAndValidateUUID(userId)
                 .flatMap(userUuid ->
@@ -37,6 +39,4 @@ public class FavoriteServiceImpl implements FavoriteService {
                                 .collect(Collectors.toSet())           // collect all into a Set<UUID>
                 );
     }
-
-
 }
