+++ itachallenge-user/src/test/java/com/itachallenge/user/controller/UserControllerTest.java
@@ -6,6 +6,7 @@ import com.itachallenge.user.dto.UserSolutionResponseDto;
 import com.itachallenge.user.exception.BadUUIDException;
 import com.itachallenge.user.exception.NotFoundException;
 import com.itachallenge.user.service.IUserSolutionService;
+import com.itachallenge.userinteraction.service.favorite.FavoriteService;
 import com.itachallenge.user.exception.UserGlobalExceptionHandler;
 import com.itachallenge.user.service.UserService;
 import org.junit.jupiter.api.*;
@@ -27,6 +28,9 @@ class UserControllerTest {
 
     @Mock
     private UserService userService;
+
+    @Mock
+    private FavoriteService favoriteService;
 
     @Mock
     private IUserSolutionService userSolutionService;
@@ -112,7 +116,7 @@ class UserControllerTest {
     void addToFavorites_WhenAdded_Returns201() {
         String userId = UUID.randomUUID().toString();
         String challengeId = UUID.randomUUID().toString();
-        when(userService.addChallengeToFavorites(userId, challengeId))
+        when(favoriteService.addChallengeToFavorites(userId, challengeId))
                 .thenReturn(Mono.just(true));
 
         webTestClient.post()
@@ -121,7 +125,7 @@ class UserControllerTest {
                 .expectStatus().isEqualTo(HttpStatus.CREATED)
                 .expectBody(Boolean.class).isEqualTo(true);
 
-        verify(userService, times(1)).addChallengeToFavorites(userId, challengeId);
+        verify(favoriteService, times(1)).addChallengeToFavorites(userId, challengeId);
     }
 
     @Test
@@ -144,7 +148,7 @@ class UserControllerTest {
     void addToFavorites_WhenAlreadyInFavorites_Returns200() {
         String userId = UUID.randomUUID().toString();
         String challengeId = UUID.randomUUID().toString();
-        when(userService.addChallengeToFavorites(userId, challengeId))
+        when(favoriteService.addChallengeToFavorites(userId, challengeId))
                 .thenReturn(Mono.just(false));
 
         webTestClient.post()
@@ -153,7 +157,7 @@ class UserControllerTest {
                 .expectStatus().isOk()
                 .expectBody(Boolean.class).isEqualTo(false);
 
-        verify(userService, times(1)).addChallengeToFavorites(userId, challengeId);
+        verify(favoriteService, times(1)).addChallengeToFavorites(userId, challengeId);
     }
 
     @Test
@@ -176,7 +180,7 @@ class UserControllerTest {
     void addToFavorites_WhenUserNotExists_Returns404() {
         String userId = UUID.randomUUID().toString();
         String challengeId = UUID.randomUUID().toString();
-        when(userService.addChallengeToFavorites(userId, challengeId))
+        when(favoriteService.addChallengeToFavorites(userId, challengeId))
                 .thenReturn(Mono.error(new NotFoundException("User not found")));
 
         webTestClient.post()
@@ -185,7 +189,7 @@ class UserControllerTest {
                 .expectStatus().isNotFound()
                 .expectBody(String.class).isEqualTo("User not found");
 
-        verify(userService, times(1)).addChallengeToFavorites(userId, challengeId);
+        verify(favoriteService, times(1)).addChallengeToFavorites(userId, challengeId);
     }
 
     @Test
@@ -208,7 +212,7 @@ class UserControllerTest {
     void addToFavorites_WhenBadFormattedId_Returns400() {
         String userId = "invalidUuid";
         String challengeId = "invalidUUid";
-        when(userService.addChallengeToFavorites(userId, challengeId))
+        when(favoriteService.addChallengeToFavorites(userId, challengeId))
                 .thenReturn(Mono.error(new BadUUIDException("Error message")));
 
         webTestClient.post()
@@ -217,7 +221,7 @@ class UserControllerTest {
                 .expectStatus().isBadRequest()
                 .expectBody(String.class).isEqualTo("The provided IDs are not valid.");
 
-        verify(userService, times(1)).addChallengeToFavorites(userId, challengeId);
+        verify(favoriteService, times(1)).addChallengeToFavorites(userId, challengeId);
     }
 
     @Test
@@ -240,7 +244,7 @@ class UserControllerTest {
     void addToFavorites_WhenUnexpectedError_Returns500() {
         String userId = UUID.randomUUID().toString();
         String challengeId = UUID.randomUUID().toString();
-        when(userService.addChallengeToFavorites(userId, challengeId))
+        when(favoriteService.addChallengeToFavorites(userId, challengeId))
                 .thenReturn(Mono.error(new Exception()));
 
         webTestClient.post()
@@ -249,7 +253,7 @@ class UserControllerTest {
                 .expectStatus().isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR)
                 .expectBody(String.class).isEqualTo("Unexpected error happened.");
 
-        verify(userService, times(1)).addChallengeToFavorites(userId, challengeId);
+        verify(favoriteService, times(1)).addChallengeToFavorites(userId, challengeId);
     }
 
     @Test
@@ -272,7 +276,7 @@ class UserControllerTest {
     void deleteFromFavorites_WhenDeleted_Returns200() {
         String userId = UUID.randomUUID().toString();
         String challengeId = UUID.randomUUID().toString();
-        when(userService.deleteChallengeFromFavorites(userId, challengeId))
+        when(favoriteService.deleteChallengeFromFavorites(userId, challengeId))
                 .thenReturn(Mono.just(true));
 
         webTestClient.delete()
@@ -281,7 +285,7 @@ class UserControllerTest {
                 .expectStatus().isEqualTo(HttpStatus.OK)
                 .expectBody(Boolean.class).isEqualTo(true);
 
-        verify(userService, times(1)).deleteChallengeFromFavorites(userId, challengeId);
+        verify(favoriteService, times(1)).deleteChallengeFromFavorites(userId, challengeId);
     }
 
     @Test
@@ -304,7 +308,7 @@ class UserControllerTest {
     void deleteFromFavorites_WhenNotInFavorites_Returns200() {
         String userId = UUID.randomUUID().toString();
         String challengeId = UUID.randomUUID().toString();
-        when(userService.deleteChallengeFromFavorites(userId, challengeId))
+        when(favoriteService.deleteChallengeFromFavorites(userId, challengeId))
                 .thenReturn(Mono.just(false));
 
         webTestClient.delete()
@@ -313,7 +317,7 @@ class UserControllerTest {
                 .expectStatus().isOk()
                 .expectBody(Boolean.class).isEqualTo(false);
 
-        verify(userService, times(1)).deleteChallengeFromFavorites(userId, challengeId);
+        verify(favoriteService, times(1)).deleteChallengeFromFavorites(userId, challengeId);
     }
 
     @Test
@@ -336,7 +340,7 @@ class UserControllerTest {
     void deleteFromFavorites_WhenUserNotExists_Returns404() {
         String userId = UUID.randomUUID().toString();
         String challengeId = UUID.randomUUID().toString();
-        when(userService.deleteChallengeFromFavorites(userId, challengeId))
+        when(favoriteService.deleteChallengeFromFavorites(userId, challengeId))
                 .thenReturn(Mono.error(new NotFoundException("User not found")));
 
         webTestClient.delete()
@@ -345,7 +349,7 @@ class UserControllerTest {
                 .expectStatus().isEqualTo(HttpStatus.NOT_FOUND)
                 .expectBody(String.class).isEqualTo("User not found");
 
-        verify(userService, times(1)).deleteChallengeFromFavorites(userId, challengeId);
+        verify(favoriteService, times(1)).deleteChallengeFromFavorites(userId, challengeId);
     }
 
     @Test
@@ -368,7 +372,7 @@ class UserControllerTest {
     void deleteFromFavorites_WhenBadFormattedId_Returns404() {
         String userId = "invalidUuid";
         String challengeId = "invalidUUid";
-        when(userService.deleteChallengeFromFavorites(userId, challengeId))
+        when(favoriteService.deleteChallengeFromFavorites(userId, challengeId))
                 .thenReturn(Mono.error(new BadUUIDException("Error message")));
 
         webTestClient.delete()
@@ -377,7 +381,7 @@ class UserControllerTest {
                 .expectStatus().isEqualTo(HttpStatus.BAD_REQUEST)
                 .expectBody(String.class).isEqualTo("The provided IDs are not valid.");
 
-        verify(userService, times(1)).deleteChallengeFromFavorites(userId, challengeId);
+        verify(favoriteService, times(1)).deleteChallengeFromFavorites(userId, challengeId);
     }
 
     @Test
@@ -400,7 +404,7 @@ class UserControllerTest {
     void deleteFromFavorites_WhenUnexpectedError_Returns500() {
         String userId = UUID.randomUUID().toString();
         String challengeId = UUID.randomUUID().toString();
-        when(userService.deleteChallengeFromFavorites(userId, challengeId))
+        when(favoriteService.deleteChallengeFromFavorites(userId, challengeId))
                 .thenReturn(Mono.error(new Exception()));
 
         webTestClient.delete()
@@ -409,7 +413,7 @@ class UserControllerTest {
                 .expectStatus().isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR)
                 .expectBody(String.class).isEqualTo("Unexpected error happened.");
 
-        verify(userService, times(1)).deleteChallengeFromFavorites(userId, challengeId);
+        verify(favoriteService, times(1)).deleteChallengeFromFavorites(userId, challengeId);
     }
 
     @Test
@@ -434,7 +438,7 @@ class UserControllerTest {
         UUID userId = UUID.randomUUID();
         Set<UUID> expectedFavorites = Set.of(UUID.randomUUID(), UUID.randomUUID());
 
-        when(userService.getUserFavorites(userId.toString())).thenReturn(Mono.just(expectedFavorites));
+        when(favoriteService.getUserFavorites(userId.toString())).thenReturn(Mono.just(expectedFavorites));
 
         webTestClient.get()
                 .uri("/itachallenge/api/v1/user/users/{userId}/favorites", userId)
@@ -444,7 +448,7 @@ class UserControllerTest {
                 .hasSize(expectedFavorites.size())
                 .contains(expectedFavorites.toArray(new UUID[0]));
 
-        verify(userService, times(1)).getUserFavorites(userId.toString());
+        verify(favoriteService, times(1)).getUserFavorites(userId.toString());
     }
 
     @Test
@@ -452,7 +456,7 @@ class UserControllerTest {
     void getUserFavorites_returns404IfUserNotFound() {
         UUID userId = UUID.randomUUID();
 
-        when(userService.getUserFavorites(userId.toString()))
+        when(favoriteService.getUserFavorites(userId.toString()))
                 .thenReturn(Mono.error(new NotFoundException("User not found")));
 
         webTestClient.get()
@@ -461,7 +465,7 @@ class UserControllerTest {
                 .expectStatus().isNotFound()
                 .expectBody(String.class).isEqualTo("User not found");
 
-        verify(userService, times(1)).getUserFavorites(userId.toString());
+        verify(favoriteService, times(1)).getUserFavorites(userId.toString());
     }
 
     @Test
@@ -469,7 +473,7 @@ class UserControllerTest {
     void getUserFavorites_returns400IfInvalidUUID() {
         String invalidUserId = "invalid-uuid";
 
-        when(userService.getUserFavorites(invalidUserId))
+        when(favoriteService.getUserFavorites(invalidUserId))
                 .thenReturn(Mono.error(new BadUUIDException("The provided IDs are not valid.")));
 
         webTestClient.get()
@@ -478,7 +482,7 @@ class UserControllerTest {
                 .expectStatus().isBadRequest()
                 .expectBody(String.class).isEqualTo("The provided IDs are not valid.");
 
-        verify(userService, times(1)).getUserFavorites(invalidUserId);
+        verify(favoriteService, times(1)).getUserFavorites(invalidUserId);
     }
 
     @Test
@@ -486,7 +490,7 @@ class UserControllerTest {
     void getUserFavorites_returns500IfUnexpectedError() {
         UUID userId = UUID.randomUUID();
 
-        when(userService.getUserFavorites(userId.toString()))
+        when(favoriteService.getUserFavorites(userId.toString()))
                 .thenReturn(Mono.error(new RuntimeException("Unexpected error")));
 
         webTestClient.get()
@@ -495,7 +499,7 @@ class UserControllerTest {
                 .expectStatus().isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR)
                 .expectBody(String.class).isEqualTo("Unexpected error happened.");
 
-        verify(userService, times(1)).getUserFavorites(userId.toString());
+        verify(favoriteService, times(1)).getUserFavorites(userId.toString());
     }
 
     @Test
